chore: move flash step to React
Changelog-entry: chore: move flash step to React
Change-type: patch
