+++ lib/gui/app/pages/main/main.js
@@ -23,15 +23,11 @@
  */
 
 const angular = require('angular')
+const { react2angular } = require('react2angular')
 const MODULE_NAME = 'Etcher.Pages.Main'
 
-require('angular-moment')
-
 const MainPage = angular.module(MODULE_NAME, [
-  'angularMoment',
-
   require('angular-ui-router'),
-  require('angular-seconds-to-date'),
 
   require('../../components/drive-selector/drive-selector'),
   require('../../components/tooltip-modal/tooltip-modal'),
@@ -54,7 +50,9 @@ const MainPage = angular.module(MODULE_NAME, [
 
 MainPage.controller('MainController', require('./controllers/main'))
 MainPage.controller('DriveSelectionController', require('./controllers/drive-selection'))
-MainPage.controller('FlashController', require('./controllers/flash'))
+MainPage.component('flash', react2angular(require('./Flash.jsx'),
+  [ 'shouldFlashStepBeDisabled', 'lastFlashErrorCode', 'progressMessage' ],
+  [ '$timeout', '$state', 'WarningModalService', 'DriveSelectorService', 'FlashErrorModalService' ]))
 
 MainPage.config(($stateProvider) => {
   $stateProvider
