+++ lib/gui/app/pages/main/styles/_main.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-svg-icon > img[disabled] {
+img[disabled] {
   opacity: $disabled-opacity;
 }
